seperate cache seed from cache utils
