+++ common/v2/config/labels.ts
@@ -1,0 +1,16 @@
+import { AddressBook } from 'v2/types';
+
+export const addressBook: { [key in string]: AddressBook } = {
+  'f1330cce-08e2-41ce-9231-5236e6aab702': {
+    label: 'Goerli ETH Test 1',
+    address: '0xc7bfc8a6bd4e52bfe901764143abef76caf2f912',
+    notes: '',
+    network: 'Goerli'
+  },
+  '13f3cbf2-de3a-4050-a0c6-521592e4b85a': {
+    label: 'ETH Test 1',
+    address: '0xc7bfc8a6bd4e52bfe901764143abef76caf2f912',
+    notes: '',
+    network: 'Ethereum'
+  }
+};
