+++ common/v2/config/settings.ts
@@ -1,0 +1,6 @@
+export const settings = {
+  fiatCurrency: 'USD',
+  darkMode: false,
+  dashboardAccounts: [],
+  inactivityTimer: 1800000
+};
